% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comb_pval.R
\name{CombP}
\alias{CombP}
\title{Convergent evidence based on combined p-values.}
\usage{
CombP(file, weight, method = c("fisher", "z.transform", "logit"),
  na.remove = FALSE)
}
\arguments{
\item{file}{A tab-delimited text file with a minimum of 3 columns. First column should
contain gene names, second column should indicate the evidence type and third column
should contain non-negative numeric values (p-values).}

\item{weight}{A numeric vector containing weights of evidence types. For example,
sample sizes of various evidence types. If not provided, equal weight is given to all
evidence types.}

\item{method}{A character string among 'fisher', 'z.transform' or 'logit'.}

\item{na.remove}{An optional argument, defaults to FALSE. Set this argument to TRUE if
all the genes were not detected across all evidence types.}
}
\value{
If all the inputs are in the correct format as suggested, then the output will
  be a dataframe containg genes, their combined p-values and corresponding ranks.
}
\description{
\code{CombP} returns ranks of the genes based on p-values combined using
  the famous 'fisher' or 'z-transform' methods.
}
\examples{
cus.weights <- c(100,50,200,300,150,400)
input_file_P <- system.file("extdata","CombP_toydata.txt",package="GenRank")
CP_ranking <- CombP(input_file_P, method = "fisher", na.remove = TRUE)
CP_ranking_z <- CombP(input_file_P, method = "z.transform", na.remove = TRUE, weight = cus.weights)
}

