% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_CE.R
\name{ComputeCE}
\alias{ComputeCE}
\title{Convergent Evidence (CE) scores of genes.}
\usage{
ComputeCE(file, PC = c("equal", "ngene", "custom"), cust.weights = NULL)
}
\arguments{
\item{file}{A tab-delimited text file with a minimum of 2 columns. First column should
contain gene names and second column should indicate the evidence type.}

\item{PC}{A character string among 'equal', 'ngene' or 'custom' indicating the prior
credibility.}

\item{cust.weights}{An optional argument required when the PC='custom'. A numeric
vector containing weights reflecting prior credibility. Should contain as many
weights as the number of evidence types.}
}
\value{
If all the inputs are in the correct format as suggested, then the output will
  be a dataframe containing genes and their ranks based on CE scores.
}
\description{
\code{ComputeCE} returns ranks of the genes based on CE scores.
}
\examples{
input_file <- system.file("extdata","CE_RP_toydata.txt",package="GenRank")
CE_ranks <- ComputeCE(input_file,PC = "equal")
evid.weight <- c(1,1,0.8,0.8,0.5,1)
CE_ranks_cust <- ComputeCE(input_file,PC = "custom", cust.weights = evid.weight)
}

