% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_RP.R
\name{ComputeRP}
\alias{ComputeRP}
\title{Convergent evidence based on rank product method.}
\usage{
ComputeRP(file, signif.type, n.perm = 100, setseed = NULL)
}
\arguments{
\item{file}{A tab-delimited text file with a minimum of 3 columns. First column should
contain gene names, second column should indicate the evidence type and third column
should contain non-negative numeric values (e.g. p-values or effect size).}

\item{signif.type}{A vector containing letters 'L' or 'H' or both. Length of the vector
should be the same as the number of evidence types. 'L' or 'H' indicate whether the
evidence type contains a low numeric value (e.g. p-value) or a high numeric value
(e.g. effect size).}

\item{n.perm}{A number indicating number of permutations used to calculate null
density.
Defaults to 100 permutations.}

\item{setseed}{An optional argument. If provided a numeric value, sampling will be put
in a reproducible state using the setseed value as seed.}
}
\value{
If all the inputs are in the correct format as suggested, then the output will
  be a dataframe containg genes, their ranks based on RP and corresponding pfp
  (equivalent to FDR).
}
\description{
\code{ComputeRP} returns ranks of the genes based on rank product method.
}
\examples{
input_file <- system.file("extdata","CE_RP_toydata.txt",package="GenRank")
signif.val <- c('L','L','H','L','H','L')
RP_ranks <- ComputeRP(input_file, signif.type = signif.val)
RP_ranks_cust <- ComputeRP(input_file, signif.type = signif.val, n.perm=200, setseed=1234)
}

